Added extra dimension to ssta in mod_forces. Changed set_avg to calculate only current ssta and put in stenCount location.
Also realized that get_data is only serial, so we shouldn't need to
use np_reduce to calculate the regression coefficients. Changed this.

Also introduced a dummy variable SSTa in get_data for the averaged
ssta for use in the regression calculation. Note: Not sure the
dimension bounds on the variable define will work here.